Fix the tests and make SQLAlchemy 1.2.8 compatible (#219)
* (More) idiomatic code for `fetchmany`, `fetchall`

I went to look if `fetchall()` downloads th entire result in memory and found that it was coded in a very low-level C style, while Python has a perfectly good idiomatic approach for such things. So I thought I'd submit a fix :-) Take it or leave it, I won't insist.

P.S. I'm editing this from GitHub UI, so haven't run any tests. And although I'm fairly sure it should work as is, consider it formally as a demonstration :-)

* Remove unused import

* Removing unreleased thrift_sasl comment

Hello. I might be removing this prematurely, but `pip search thrift_sasl` now shows 0.3.0. Can this comment be removed?

Also, I filled out the Dropbox Contributor License.

* Add __enter__, __exit__ methods

The methods allow us to use connections and cursors with `with`s
statements.

* Use a setter for arraysize

The reason behind this is that non-integer sizes or None (or 0) do not
make sense.

Especially with None, which might be a common pattern, this should be
prevented as it raises (cryptic) errors on the Thrift size.

* Add script to generate new TCLIService files

* Add decimal/timestamp conversion

* Add the Travis pin

* Fix Flake8 violations

* Add Codecov badge